requirements.txt
streamlit
pandas
matplotlib
seaborn
numpy
gspread
gspread-dataframe
google-auth